プログラムをdockerで実行できるようにした
